Fix an issue when the clockpicker is initialised with 0:00
The clockpicker would should 0:00 PM when first opened. Even when used in 24 hour mode. When used in 24 hour mode, the PM shouldn't even be visible. In 12 hour mode it should show AM and not PM. When a user selects a time before 12:00. The PM time was saved. Say a user selected 10:00, 22:00 would be stored.